Add the Site's Components
In this commit, components were added and setup to handle different
functions needed to keep the website running smoothly.

NB: React-Router-Dom provides a good way of passing down props from the
parent to its children router by using the special prop, "context" and
then passing the desired data into the "Outlet" component which can
thereafter be accessed anywhere using the "useOutletContext" hook in its
children router, even in deep nested components. This was not
implemented in this commit and project but its an awesome way to pass
down props when needed.