Setup the Site's Pages
In this commit, the neccessary pages required for the website were setup
to handle different functions required to ensure the smooth working of
the site.

NB: One of the important and salient point to note is that "redirect"
functions are only to be used in the action and loader functions.